producer/requirements.txt
kafka-python
requests
python-dotenv
boto3
